requirements.txt
pandas
streamlit
pyDomo
pyodbc
pymssql
python-decouple


npm run dump streamlit_app  python-decouple pyodbc pyDomo 


pip install -U hydralit_component